# perf(core): implement parallel file processing for 74% performance improvement