## TLDR

Implement parallel file processing in ReadManyFilesTool using Promise.allSettled to replace sequential for-loop, achieving **74% performance
improvement** (408ms → 107ms for batch operations). Builds on Han5991's async file detection work (#3286) to enable concurrent file type detection
and content reading without breaking existing functionality.

## Dive Deeper

### Problem
The current implementation processes files sequentially using a for-loop with await, causing unnecessary waiting time when processing multiple files.
Each file operation (type detection + content reading) blocked the next file from starting, leading to poor performance on large projects.

### Solution
- **Replace sequential for-loop** with `Promise.allSettled()` for concurrent file processing
- **Maintain error isolation**: Individual file failures don't affect other files
- **Preserve all existing functionality**: Error handling, skipped file tracking, and content formatting remain identical
- **Leverage existing async infrastructure**: Built on Han5991's async `detectFileType` and `processSingleFileContent` functions

### Technical Implementation
```typescript
  // Before: Sequential processing
  for (const filePath of sortedFiles) {
    const fileType = await detectFileType(filePath);
    const content = await processSingleFileContent(filePath, ...);
  }

  // After: Parallel processing
  const promises = sortedFiles.map(async (filePath) => {
    const fileType = await detectFileType(filePath);
    const content = await processSingleFileContent(filePath, ...);
    return { filePath, fileType, content };
  });
  const results = await Promise.allSettled(promises);
```

###Performance Results

- 4 files: 408ms → 107ms (74% improvement)
- Scaling: Performance gap increases with more files
- Memory: Minimal overhead increase
- Reliability: All existing error handling preserved

## Reviewer Test Plan

Performance Verification

1. Run existing tests: npm test -- --run src/tools/read-many-files.test.ts
    - All 25 tests should pass (22 existing + 3 new)
    - New tests specifically verify parallel processing performance and concurrency
2. Manual testing with large file sets:

### Create test scenario
  mkdir -p test-batch && cd test-batch
  for i in {1..20}; do echo "Content $i" > file$i.txt; done

### Test the CLI
  gemini "read all txt files in this directory"
  4. Performance comparison prompts:
    - "Show me all TypeScript files in src/" (should be noticeably faster)
    - "Read all package.json files in the project" (multi-file processing)

  Concurrency Verification

  The test suite includes mocked timing tests that verify:
  - All file operations start simultaneously (not sequentially)
  - Total processing time is ~max(individual_time) rather than sum(individual_times)
  - Execution order patterns confirm parallel behavior

  Error Handling Validation

  - Mixed success/failure: Create some unreadable files to verify partial success handling
  - All failures: Test with non-existent files to ensure graceful degradation
  - Memory pressure: Test with many files to verify no resource leaks

## Testing Matrix

  |          | 🍏  | 🪟  | 🐧  |
  |----------|-----|-----|-----|
  | npm run  | ✅   | ❓   | ❓   |
  | npx      | ❓   | ❓   | ❓   |
  | Docker   | ❓   | ❓   | ❓   |
  | Podman   | ❓   | -   | -   |
  | Seatbelt | ❓   | -   | -   |

  Tested on macOS with:
  - npm run preflight - All tests, linting, and type checks pass
  - npm test - Full test suite passes (25/25 tests)
  - Manual CLI testing with various file sets

## Linked issues / bugs
- Resolves #4712

Builds on:
  - Related to #3286 (Han5991's async file detection implementation)

Performance improvement for:
  - Large project file reading operations
  - Batch file processing scenarios
  - Multi-file content analysis workflows

  This PR directly addresses performance bottlenecks in file-heavy operations while maintaining full backward compatibility and error handling
  robustness.